+++ src/fst/code.py
@@ -1620,6 +1620,9 @@ def _coerce_to_withitem(code: Code, parse_params: Mapping[str, Any] = {}, *, san
 
     fst_ = code_as_expr(code, parse_params, sanitize=sanitize, coerce=True)
 
+    if fst_.a.__class__ is Starred:
+        raise NodeError('Starred not allowed in withitem')
+
     if fst_.is_parenthesized_tuple() is False:
         fst_._delimit_node()