+++ src/fst/fst_misc.py
@@ -251,7 +251,7 @@ _DEFAULT_AST_FIELD = {kls: field for field, classes in [  # builds to {Module: '
 ] for kls in classes}
 
 _ASTS_LEAF_TUPLE_OR_MATCHSEQ = frozenset([Tuple, MatchSequence])
-_ASTS_LEAF_SINGLE_CHAR_DELIM_SEQ = frozenset([Tuple, List, Set, MatchSequence])
+# _ASTS_LEAF_SINGLE_CHAR_DELIM_SEQ = frozenset([Tuple, List, Set, MatchSequence])
 
 _re_dump_line_tail     = re.compile(r'\s* ( \#.*$ | \\$ | ; (?: \s* (?: \#.*$ | \\$ ) )? )', re.VERBOSE)
 _re_one_space_or_end   = re.compile(r'\s|$')
@@ -2094,7 +2094,7 @@ def _undelimit_node(self: fst.FST, field: str = 'elts') -> bool:
     - `bool`: Whether delimiters were removed or not (they may not be for an empty tuple).
     """
 
-    assert self.a.__class__ in _ASTS_LEAF_SINGLE_CHAR_DELIM_SEQ  # _ASTS_LEAF_TUPLE_OR_MATCHSEQ
+    assert self.a.__class__ in _ASTS_LEAF_TUPLE_OR_MATCHSEQ  # _ASTS_LEAF_SINGLE_CHAR_DELIM_SEQ
 
     if not (body := getattr(self.a, field, None)):
         return False